(WIP)pybtp: Make the module pip-installable
pybtp can be reused in other non-pts projects.
The module could be installed with pip:
python -m build
pip install -U .
