+++ autopts/pybtp/__init__.py
@@ -14,3 +14,6 @@
 #
 
 """BTP"""
+
+from btp import *
+# from .btp.btpdefs import *
