+++ autopts/pybtp/btp/__init__.py
@@ -15,31 +15,32 @@
 
 """BTP"""
 
-from autopts.pybtp.btp.btp import *
-from autopts.pybtp.btp.gap import *
-from autopts.pybtp.btp.gatt import *
-from autopts.pybtp.btp.gatt_cl import *
-from autopts.pybtp.btp.l2cap import *
-from autopts.pybtp.btp.mesh import *
-from autopts.pybtp.btp.mmdl import *
-from autopts.pybtp.btp.sm import *
-from autopts.pybtp.btp.vcs import *
-from autopts.pybtp.btp.vocs import *
-from autopts.pybtp.btp.aics import *
-from autopts.pybtp.btp.pacs import *
-from autopts.pybtp.btp.ascs import *
-from autopts.pybtp.btp.bap import *
-from autopts.pybtp.btp.has import *
-from autopts.pybtp.btp.core import *
-from autopts.pybtp.btp.csis import *
-from autopts.pybtp.btp.micp import *
-from autopts.pybtp.btp.mics import *
-from autopts.pybtp.btp.ccp import *
-from autopts.pybtp.btp.vcp import *
-from autopts.pybtp.btp.cas import *
-from autopts.pybtp.btp.mcp import *
-from autopts.pybtp.btp.gmcs import *
-from autopts.pybtp.btp.hap import *
-from autopts.pybtp.btp.cap import *
-from autopts.pybtp.btp.csip import *
+from .btpdefs import *
+from .btp import *
+from .gap import *
+from .gatt import *
+from .gatt_cl import *
+from .l2cap import *
+from .mesh import *
+from .mmdl import *
+from .sm import *
+from .vcs import *
+from .vocs import *
+from .aics import *
+from .pacs import *
+from .ascs import *
+from .bap import *
+from .has import *
+from .core import *
+from .csis import *
+from .micp import *
+from .mics import *
+from .ccp import *
+from .vcp import *
+from .cas import *
+from .mcp import *
+from .gmcs import *
+from .hap import *
+from .cap import *
+from .csip import *
 # GENERATOR append 1
