+++ autopts/pybtp/btp/btp.py
@@ -15,29 +15,75 @@
 
 """Wrapper around btp messages. The functions are added as needed."""
 
-from collections import namedtuple
 from uuid import UUID
 import logging
-import re
 import struct
 
-from autopts.ptsprojects.stack import get_stack
-from autopts.ptsprojects.testcase import MMI
-from .. import defs
-from autopts.pybtp.types import BTPError, att_rsp_str
-
-#  get IUT global method from iutctl
-get_iut = None
-
-# Address
-LeAddress = namedtuple('LeAddress', 'addr_type addr')
-PTS_BD_ADDR = LeAddress(addr_type=0, addr='000000000000')
-LT2_BD_ADDR = LeAddress(addr_type=0, addr='000000000000')
-
-# Devices found
-LeAdv = namedtuple('LeAdv', 'addr_type addr rssi flags eir')
+from .btpdefs import defs
+from .btpdefs.types import BTPError, att_rsp_str, addr2btp_ba
 
 CONTROLLER_INDEX = 0
+
+
+def raise_stack_getter_exception():
+    raise Exception('stack_getter() has not been set!')
+
+
+# Depending on the project, the Stack instance can
+# be implemented as global one or each Tester thread
+# can have its own Stack instance.
+stack_getter = raise_stack_getter_exception
+
+
+def get_stack():
+    return stack_getter()
+
+
+def set_stack_getter(_stack_getter):
+    global stack_getter
+
+    stack_getter = _stack_getter
+
+
+def address_to_ba(bd_addr_type=None, bd_addr=None):
+    """Get Bluetooth address in bytearray format"""
+    stack = get_stack()
+    data = bytearray()
+    bd_addr_ba = addr2btp_ba(stack.pts_addr_get(bd_addr))
+    bd_addr_type_ba = chr(stack.pts_addr_type_get(bd_addr_type)).encode('utf-8')
+    data.extend(bd_addr_type_ba)
+    data.extend(bd_addr_ba)
+    return data
+
+
+def pts_addr_type_get(bd_addr=None):
+    return get_stack().pts_addr_type_get(bd_addr=bd_addr, lt=1)
+
+
+def pts_addr_get(bd_addr_type=None):
+    return get_stack().pts_addr_get(bd_addr_type=bd_addr_type, lt=1)
+
+
+def lt2_addr_type_get(bd_addr_type=None):
+    return get_stack().pts_addr_type_get(bd_addr_type=bd_addr_type, lt=2)
+
+
+def lt2_addr_get(bd_addr=None):
+    return get_stack().pts_addr_get(bd_addr=bd_addr, lt=2)
+
+
+def btp_hdr_check(rcv_hdr, exp_svc_id, exp_op=None):
+    if rcv_hdr.svc_id != exp_svc_id:
+        raise BTPError("Incorrect service ID %s in the response, expected %s!"
+                       % (rcv_hdr.svc_id, exp_svc_id))
+
+    if rcv_hdr.op == defs.BTP_STATUS:
+        raise BTPError("Error opcode in response!")
+
+    if exp_op and exp_op != rcv_hdr.op:
+        raise BTPError(
+            "Invalid opcode 0x%.2x in the response, expected 0x%.2x!" %
+            (rcv_hdr.op, exp_op))
 
 
 def read_supp_svcs():
@@ -84,13 +130,13 @@ CORE = {
     "gatt_cl_unreg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_UNREGISTER_SERVICE,
                       defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_GATTC),
     "vcs_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
-                 defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_VCS),
+                defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_VCS),
     "vocs_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
                  defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_VOCS),
     "aics_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
                  defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_AICS),
     "ias_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
-                 defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_IAS),
+                defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_IAS),
     "pacs_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
                  defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_PACS),
     "ascs_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
@@ -114,7 +160,7 @@ CORE = {
     "mcp_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
                 defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_MCP),
     "gmcs_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
-                defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_GMCS),
+                 defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_GMCS),
     "hap_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
                 defs.BTP_INDEX_NONE, defs.BTP_SERVICE_ID_HAP),
     "cap_reg": (defs.BTP_SERVICE_ID_CORE, defs.CORE_REGISTER_SERVICE,
@@ -133,235 +179,6 @@ CORE = {
 }
 
 
-def clear_verify_values():
-    stack = get_stack()
-    stack.gatt_cl.verify_values = []
-
-
-def add_to_verify_values(item):
-    stack = get_stack()
-    stack.gatt_cl.verify_values.append(item)
-
-
-def get_verify_values():
-    stack = get_stack()
-    return stack.gatt_cl.verify_values
-
-
-def extend_verify_values(item):
-    stack = get_stack()
-    stack.gatt_cl.verify_values.extend(item)
-
-
-def verify_att_error(description):
-    logging.debug("description=%r", description)
-
-    description_values = []
-
-    for err_code, err_string in att_rsp_str.items():
-        if err_string and err_string in description:
-            description_values.append(err_string)
-
-    verify_values = get_verify_values()
-    logging.debug("Verifying values: %r", verify_values)
-
-    logging.debug("Description values: %r", description_values)
-
-    for value in description_values:
-        logging.debug("Verifying: %r", value)
-
-        try:
-            if value not in verify_values:
-                logging.debug("Verification failed, value not in verify values")
-                return False
-        except TypeError:
-            logging.debug("Value under verification is not string")
-
-    logging.debug("All verifications passed")
-
-    clear_verify_values()
-
-    return True
-
-
-def verify_description(description):
-    """A function to verify that values are in PTS MMI description
-
-    Returns True if verification is successful, False if not.
-
-    description -- MMI description
-
-    """
-    logging.debug("description=%r", description)
-
-    description_values = re.findall(r"(?:'|=\s+)([0-9-xA-Fa-f]{2,})", description)
-    logging.debug("Description values: %r", description_values)
-
-    verify_values = get_verify_values()
-    logging.debug("Verifying values: %r", verify_values)
-
-    # verify_values shall not be a string: all its characters will be verified
-    assert isinstance(verify_values, list), "verify_values should be a list!"
-
-    converted_verify = []
-
-    # convert small values to int to simplify verification.
-    # Some values are displayed with leading zeros
-    for verify in verify_values:
-        verify = verify.upper()
-        if len(verify) < 8:
-            converted_verify.append(int(verify, 16))
-        else:
-            converted_verify.append(verify)
-
-    for value in description_values:
-        logging.debug("Verifying: %r", value)
-
-        value = value.upper()
-
-        if len(value) < 8:
-            value = int(value, 16)
-
-        try:
-            if value not in converted_verify:
-                logging.debug("Verification failed, value not in verify values")
-                return False
-        except TypeError:
-            logging.debug("Value under verification is not string")
-
-    logging.debug("All verifications passed")
-
-    clear_verify_values()
-
-    return True
-
-
-def verify_description_truncated(description):
-    """A function to verify that truncated values are in PTS MMI description.
-
-    Verification is successful if the PTS MMI description contains a value
-    starting with the value under verification.
-
-    Returns True if verification is successful, False if not.
-
-    description -- MMI description
-
-    """
-    logging.debug("description=%r", description)
-
-    description_values = re.findall(r"(?:'|=\s+)([0-9-xA-Fa-f]{2,})", description)
-    logging.debug("Description values: %r", description_values)
-
-    verify_values = get_verify_values()
-    logging.debug("Verifying values: %r", verify_values)
-
-    # verify_values shall not be a string: all its characters will be verified
-    assert isinstance(verify_values, list), "verify_values should be a list!"
-
-    verify_values = list(map(str.upper, verify_values))
-    description_values = list(map(str.upper, description_values))
-
-    unverified_desc = [x for x in description_values if x not in verify_values]
-    if unverified_desc:
-        logging.debug("Verifying for partial matches: %r", unverified_desc)
-
-        for desc_value in unverified_desc:
-            matches = [x for x in verify_values if desc_value.startswith(x) and len(x) > 8]
-            if not matches:
-                logging.debug("Verification failed, %r not in verify values", desc_value)
-                return False
-
-    logging.debug("All verifications passed")
-
-    clear_verify_values()
-
-    return True
-
-
-def verify_multiple_read_description(description):
-    """A function to verify that merged multiple read att values are in
-
-    PTS MMI description.
-
-    Returns True if verification is successful, False if not.
-
-    description -- MMI description
-
-    """
-    logging.debug("description=%r", description)
-
-    MMI.reset()
-    MMI.parse_description(description)
-    description_values = MMI.args
-    logging.debug("Description values: %r", description_values)
-
-    got_mtp_read = [''.join(description_values)]
-
-    verify_values = get_verify_values()
-    logging.debug("Verifying values: %r", verify_values)
-
-    # verify_values shall not be a string: all its characters will be verified
-    assert isinstance(verify_values, list), "verify_values should be a list!"
-
-    exp_mtp_read = ""
-    for value in verify_values:
-        try:
-            exp_mtp_read = exp_mtp_read.join(value)
-        except TypeError:
-            value = value.decode("utf-8")
-            exp_mtp_read = exp_mtp_read.join(value)
-
-    if exp_mtp_read not in got_mtp_read:
-        logging.debug("Verification failed, value not in description")
-        return False
-
-    logging.debug("Multiple read verifications passed")
-
-    clear_verify_values()
-
-    return True
-
-
-def parse_passkey_description(description):
-    """A function to parse passkey from description
-
-    PTS MMI description.
-
-    Returns passkey if successful, None if not.
-
-    description -- MMI description
-    """
-    logging.debug("description=%r", description)
-
-    match = re.search(r"\b[0-9]+\b", description)
-    if match:
-        pk = match.group(0)
-        logging.debug("passkey=%r", pk)
-        return int(pk)
-
-    return None
-
-
-def parse_handle_description(description):
-    """A function to parse handle from description
-
-    PTS MMI description.
-
-    Returns passkey if successful, None if not.
-
-    description -- MMI description
-    """
-    logging.debug("description=%r", description)
-
-    match = re.search(r"\bhandle (?:0x)?([0-9A-Fa-f]+)\b", description)
-    if match:
-        handle = match.group(1)
-        logging.debug("handle=%r", handle)
-        return int(handle)
-
-    return None
-
-
 def btp_hdr_check(rcv_hdr, exp_svc_id, exp_op=None):
     if rcv_hdr.svc_id != exp_svc_id:
         raise BTPError("Incorrect service ID %s in the response, expected %s!"
@@ -374,57 +191,6 @@ def btp_hdr_check(rcv_hdr, exp_svc_id, exp_op=None):
         raise BTPError(
             "Invalid opcode 0x%.2x in the response, expected 0x%.2x!" %
             (rcv_hdr.op, exp_op))
-
-
-def bd_addr_convert(bdaddr):
-    """ Remove colons from address and convert to lower case """
-    if isinstance(bdaddr, bytes):
-        bdaddr = bdaddr.decode("utf-8")
-    return "".join(bdaddr.split(':')).lower()
-
-
-def pts_addr_get(bd_addr=None):
-    """" If address provided, convert, otherwise, use stored address. """
-    if bd_addr is None:
-        return PTS_BD_ADDR.addr
-    return bd_addr_convert(bd_addr)
-
-
-def pts_addr_type_get(bd_addr_type=None):
-    """"
-    If address type provided, return it, otherwise,
-    use stored address.
-    """
-    if bd_addr_type is None:
-        return PTS_BD_ADDR.addr_type
-    return bd_addr_type
-
-
-def set_pts_addr(addr, addr_type):
-    global PTS_BD_ADDR
-    PTS_BD_ADDR = LeAddress(addr_type=addr_type, addr=bd_addr_convert(addr))
-
-
-def lt2_addr_get(bd_addr=None):
-    """" If address provided, convert, otherwise, use stored address. """
-    if bd_addr is None:
-        return LT2_BD_ADDR.addr
-    return bd_addr_convert(bd_addr)
-
-
-def lt2_addr_type_get(bd_addr_type=None):
-    """"
-    If address type provided, return it, otherwise,
-    use stored address.
-    """
-    if bd_addr_type is None:
-        return LT2_BD_ADDR.addr_type
-    return bd_addr_type
-
-
-def set_lt2_addr(addr, addr_type):
-    global LT2_BD_ADDR
-    LT2_BD_ADDR = LeAddress(addr_type=addr_type, addr=bd_addr_convert(addr))
 
 
 def core_reg_svc_gap():
@@ -531,17 +297,20 @@ def core_reg_svc_vcs():
     iutctl = get_iut()
     iutctl.btp_socket.send_wait_rsp(*CORE['vcs_reg'])
 
+
 def core_reg_svc_vocs():
     logging.debug("%s", core_reg_svc_vocs.__name__)
 
     iutctl = get_iut()
     iutctl.btp_socket.send_wait_rsp(*CORE['vocs_reg'])
 
+
 def core_reg_svc_aics():
     logging.debug("%s", core_reg_svc_aics.__name__)
 
     iutctl = get_iut()
     iutctl.btp_socket.send_wait_rsp(*CORE['aics_reg'])
+
 
 def core_reg_svc_ias():
     logging.debug("%s", core_reg_svc_ias.__name__)
@@ -632,6 +401,7 @@ def core_reg_svc_gmcs():
 
     iutctl = get_iut()
     iutctl.btp_socket.send_wait_rsp(*CORE['gmcs_reg'])
+
 
 def core_reg_svc_hap():
     logging.debug("%s", core_reg_svc_hap.__name__)
@@ -727,7 +497,6 @@ def init(get_iut_method):
     global get_iut
 
     get_iut = get_iut_method
-    set_event_handler(event_handler)
 
 
 from .gap import GAP_EV
@@ -753,8 +522,6 @@ from .hap import HAP_EV
 from .cap import CAP_EV
 from .csip import CSIP_EV
 # GENERATOR append 2
-
-from autopts.pybtp.iutctl_common import set_event_handler
 
 
 def event_handler(hdr, data):
