+++ autopts/pybtp/btp/btpdefs/gatt.py
@@ -1,0 +1,60 @@
+#
+# auto-pts - The Bluetooth PTS Automation Framework
+#
+# Copyright (c) 2017, Intel Corporation.
+# Copyright (c) 2024, Codecoup.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms and conditions of the GNU General Public License,
+# version 2, as published by the Free Software Foundation.
+#
+# This program is distributed in the hope it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+from threading import Event
+
+
+class GattAttribute:
+    def __init__(self, handle, perm, uuid, att_rsp):
+        self.handle = handle
+        self.perm = perm
+        self.uuid = uuid
+        self.att_read_rsp = att_rsp
+
+
+class GattService(GattAttribute):
+    def __init__(self, handle, perm, uuid, att_rsp, end_handle=None):
+        super().__init__(handle, perm, uuid, att_rsp)
+        self.end_handle = end_handle
+
+
+class GattPrimary(GattService):
+    pass
+
+
+class GattSecondary(GattService):
+    pass
+
+
+class GattServiceIncluded(GattAttribute):
+    def __init__(self, handle, perm, uuid, att_rsp, incl_svc_hdl, end_grp_hdl):
+        GattAttribute.__init__(self, handle, perm, uuid, att_rsp)
+        self.incl_svc_hdl = incl_svc_hdl
+        self.end_grp_hdl = end_grp_hdl
+
+
+class GattCharacteristic(GattAttribute):
+    def __init__(self, handle, perm, uuid, att_rsp, prop, value_handle):
+        GattAttribute.__init__(self, handle, perm, uuid, att_rsp)
+        self.prop = prop
+        self.value_handle = value_handle
+
+
+class GattCharacteristicDescriptor(GattAttribute):
+    def __init__(self, handle, perm, uuid, att_rsp, value):
+        GattAttribute.__init__(self, handle, perm, uuid, att_rsp)
+        self.value = value
+        self.has_changed_cnt = 0
+        self.has_changed = Event()
