+++ autopts/pybtp/btp/cap.py
@@ -18,11 +18,11 @@ import binascii
 import logging
 import struct
 
-from autopts.pybtp import defs
-from autopts.pybtp.btp.btp import CONTROLLER_INDEX, get_iut_method as get_iut,\
-    btp_hdr_check, pts_addr_get, pts_addr_type_get
-from autopts.pybtp.btp.gap import __gap_current_settings_update
-from autopts.pybtp.types import addr2btp_ba, BTPError
+from .btpdefs import defs
+from .btp import CONTROLLER_INDEX, get_iut_method as get_iut, \
+    btp_hdr_check, address_to_ba
+from .gap import __gap_current_settings_update
+from .btpdefs.types import BTPError
 
 CAP = {
     'read_supported_cmds': (defs.BTP_SERVICE_ID_CAP,
@@ -65,15 +65,6 @@ def cap_command_rsp_succ(timeout=20.0):
     btp_hdr_check(tuple_hdr, defs.BTP_SERVICE_ID_CAP)
 
     return tuple_data
-
-
-def address_to_ba(bd_addr_type=None, bd_addr=None):
-    data = bytearray()
-    bd_addr_ba = addr2btp_ba(pts_addr_get(bd_addr))
-    bd_addr_type_ba = chr(pts_addr_type_get(bd_addr_type)).encode('utf-8')
-    data.extend(bd_addr_type_ba)
-    data.extend(bd_addr_ba)
-    return data
 
 
 def cap_discover(bd_addr_type=None, bd_addr=None):
