+++ autopts/pybtp/btp/csip.py
@@ -16,10 +16,10 @@
 import logging
 import struct
 
-from autopts.pybtp import defs
-from autopts.pybtp.btp.btp import CONTROLLER_INDEX, get_iut_method as get_iut, \
-    btp_hdr_check, pts_addr_get, pts_addr_type_get
-from autopts.pybtp.types import addr2btp_ba
+from .btpdefs import defs
+from .btp import CONTROLLER_INDEX, get_iut_method as get_iut, \
+    btp_hdr_check, address_to_ba
+
 
 log = logging.debug
 
@@ -34,15 +34,6 @@ CSIP = {
                              defs.CSIP_START_ORDERED_ACCESS,
                              CONTROLLER_INDEX),
 }
-
-
-def address_to_ba(bd_addr_type=None, bd_addr=None):
-    data = bytearray()
-    bd_addr_ba = addr2btp_ba(pts_addr_get(bd_addr))
-    bd_addr_type_ba = chr(pts_addr_type_get(bd_addr_type)).encode('utf-8')
-    data.extend(bd_addr_type_ba)
-    data.extend(bd_addr_ba)
-    return data
 
 
 def csip_command_rsp_succ(timeout=20.0):
