+++ autopts/pybtp/btp/gmcs.py
@@ -14,14 +14,12 @@
 #
 
 """Wrapper around btp messages. The functions are added as needed."""
-import binascii
 import logging
 import struct
 
-from autopts.pybtp import defs
-from autopts.pybtp.btp.btp import CONTROLLER_INDEX, get_iut_method as get_iut, btp_hdr_check, pts_addr_get, \
-    pts_addr_type_get
-from autopts.pybtp.types import BTPError, addr2btp_ba
+from .btpdefs import defs
+from .btp import CONTROLLER_INDEX, get_iut_method as get_iut, btp_hdr_check
+from .btpdefs.types import BTPError
 
 GMCS = {
     'read_supported_cmds':      (defs.BTP_SERVICE_ID_GMCS,
@@ -55,15 +53,6 @@ def gmcs_command_rsp_succ(timeout=20.0):
 
     btp_hdr_check(tuple_hdr, defs.BTP_SERVICE_ID_GMCS)
     return tuple_data
-
-
-def address_to_ba(bd_addr_type=None, bd_addr=None):
-    data = bytearray()
-    bd_addr_ba = addr2btp_ba(pts_addr_get(bd_addr))
-    bd_addr_type_ba = chr(pts_addr_type_get(bd_addr_type)).encode('utf-8')
-    data.extend(bd_addr_type_ba)
-    data.extend(bd_addr_ba)
-    return data
 
 
 def gmcs_control_point_cmd(opcode, use_param, param=None):
