+++ autopts/pybtp/btp/ias.py
@@ -15,9 +15,9 @@
 
 import logging
 
-from autopts.pybtp import defs
-from autopts.ptsprojects.stack import get_stack
-from autopts.pybtp.btp.btp import CONTROLLER_INDEX, get_iut_method as get_iut
+from .btpdefs import defs
+from .btp import get_stack
+
 
 def ias_ev_out_alert_action(ias, data, data_len):
     logging.debug("%s %r", ias_ev_out_alert_action.__name__, data)
