+++ autopts/pybtp/btp/vcp.py
@@ -18,10 +18,9 @@ import binascii
 import logging
 import struct
 
-from autopts.pybtp import defs
-from autopts.pybtp.btp.btp import CONTROLLER_INDEX, get_iut_method as get_iut, btp_hdr_check, pts_addr_get, \
-    pts_addr_type_get
-from autopts.pybtp.types import BTPError, addr2btp_ba
+from .btpdefs import defs
+from .btp import CONTROLLER_INDEX, get_iut_method as get_iut, btp_hdr_check, address_to_ba
+from .btpdefs.types import BTPError
 
 VCP = {
     'read_supported_cmds': (defs.BTP_SERVICE_ID_VCP,
@@ -70,15 +69,6 @@ def vcp_command_rsp_succ(timeout=20.0):
 
     btp_hdr_check(tuple_hdr, defs.BTP_SERVICE_ID_VCP)
     return tuple_data
-
-
-def address_to_ba(bd_addr_type=None, bd_addr=None):
-    data = bytearray()
-    bd_addr_ba = addr2btp_ba(pts_addr_get(bd_addr))
-    bd_addr_type_ba = chr(pts_addr_type_get(bd_addr_type)).encode('utf-8')
-    data.extend(bd_addr_type_ba)
-    data.extend(bd_addr_ba)
-    return data
 
 
 def vcp_discover(bd_addr_type=None, bd_addr=None):
