+++ autopts/pybtp/btp/vocs.py
@@ -19,12 +19,9 @@ import binascii
 import logging
 import struct
 
-from autopts.pybtp import defs
-from autopts.ptsprojects.stack import get_stack
-from autopts.pybtp.types import BTPError, addr2btp_ba
-from autopts.pybtp.btp.btp import CONTROLLER_INDEX, btp_hdr_check, get_iut_method as get_iut, pts_addr_get, \
-    pts_addr_type_get
-
+from .btpdefs import defs
+from .btpdefs.types import BTPError
+from .btp import CONTROLLER_INDEX, btp_hdr_check, get_iut_method as get_iut, address_to_ba
 
 VOCS = {
     'read_supported_cmds': (defs.BTP_SERVICE_ID_VCP,
@@ -57,15 +54,6 @@ def vocs_command_rsp_succ(op=None):
     btp_hdr_check(tuple_hdr, defs.BTP_SERVICE_ID_VOCS, op)
 
     return tuple_data
-
-
-def address_to_ba(bd_addr_type=None, bd_addr=None):
-    data = bytearray()
-    bd_addr_ba = addr2btp_ba(pts_addr_get(bd_addr))
-    bd_addr_type_ba = chr(pts_addr_type_get(bd_addr_type)).encode('utf-8')
-    data.extend(bd_addr_type_ba)
-    data.extend(bd_addr_ba)
-    return data
 
 
 def vocs_audio_desc(string):
