+++ setup.py
@@ -3,8 +3,27 @@ import setuptools
 with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
+# setuptools.setup(
+#     name="autopts",
+#     version="0.0.1",
+#     author="Example Author",
+#     author_email="author@example.com",
+#     description="A small example package",
+#     long_description=long_description,
+#     long_description_content_type="text/markdown",
+#     url="https://github.com/pypa/sampleproject",
+#     classifiers=[
+#         "Programming Language :: Python :: 3",
+#         "License :: OSI Approved :: MIT License",
+#         "Operating System :: OS Independent",
+#     ],
+#     # package_dir={"": "src"},
+#     packages=setuptools.find_packages(include=['autopts']),
+#     python_requires=">=3.6",
+# )
+
 setuptools.setup(
-    name="autopts",
+    name="pybtp",
     version="0.0.1",
     author="Example Author",
     author_email="author@example.com",
@@ -17,7 +36,10 @@ setuptools.setup(
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
-    # package_dir={"": "src"},
-    packages=setuptools.find_packages(include=['autopts']),
+    package_dir={"": "./autopts/pybtp"},
+    # packages=setuptools.find_packages(
+    #     where='./autopts',
+    #     include=['pybtp']
+    # ),
     python_requires=">=3.6",
 )
